OMTD: Calculate image size in chronos runner
Before dispatching something to chronos, ask docker registry API
for information about the compressed image size. If the size is
above some threashold (default: 1G), set the disk requirements
of the task to a high value (default: 50G), otherwise to a low
value (default: 256M).

This change will help the cluster to decide where each task
should be executed and it is OpenMinTeD specific.